Address various review feedbacks
1) Fix style of python notebook;
2) GraphStore:
   - improve a corner case by avoiding full set iteration;
   - add some null checking;
   - move utilities around.
